Refactors and bug fixes
