+++ app/onboarding/login.tsx
@@ -1,3 +1,4 @@
+import { router } from "expo-router"
 import { useState } from "react"
 import {
   ImageBackground,
@@ -6,26 +7,23 @@ import {
   TouchableOpacity,
   View
 } from "react-native"
-import image from "../assets/backgrounds/login.jpg"
-import GlassCard from "../components/containers/GlassCard"
-import EmailIcon from "../components/icons/EmailIcon"
-import LockIcon from "../components/icons/LockIcon"
-import Entry from "../components/inputs/Entry"
-import { styles } from "../styles"
-import { Link, router } from "expo-router"
-import LongButton from "../components/inputs/LongButton"
-
+import image from "../../assets/backgrounds/login.jpg"
+import GlassCard from "../../components/containers/GlassCard"
+import EmailIcon from "../../components/icons/EmailIcon"
+import LockIcon from "../../components/icons/LockIcon"
+import Entry from "../../components/inputs/Entry"
+import LongButton from "../../components/inputs/LongButton"
+import { styles } from "../../styles"
 
 export default function LoginPage() {
   const [emailInput, setEmailInput] = useState<string>("")
   const [rawPasswordInput, setRawPasswordInput] = useState<string>("")
 
-
   /**
    * Hanlder untuk tombol register.
    */
   const handleRegister = () => {
-    router.push("/register")
+    router.push("/onboarding/register")
   }
 
   /**
@@ -34,7 +32,6 @@ export default function LoginPage() {
   const handleLogin = () => {
     // TODO: Implement login!
   }
-
 
   return (
     <ImageBackground source={image} blurRadius={4} className="h-full -z-10">
@@ -48,19 +45,21 @@ export default function LoginPage() {
             placeholder="Masukkan email"
             icon={EmailIcon}
             value={emailInput}
-            setValue={setEmailInput}/>
+            setValue={setEmailInput}
+          />
           <Entry
             placeholder="Masukkan kata sandi"
             icon={LockIcon}
             value={rawPasswordInput}
             setValue={setRawPasswordInput}
-            type="password" />
+            type="password"
+          />
           <TouchableOpacity activeOpacity={0.5} onPress={handleRegister}>
             <Text className=" text-white text-caption -mt-[4px] mb-[6px] mx-[2px]">
               Belum punya akun? Klik disini
             </Text>
           </TouchableOpacity>
-          <LongButton label="Masuk" onClick={handleLogin}/>
+          <LongButton label="Masuk" onClick={handleLogin} />
         </GlassCard>
       </SafeAreaView>
     </ImageBackground>
