+++ components/inputs/BackButton.tsx
@@ -1,5 +1,6 @@
-import { Text, TouchableOpacity } from "react-native"
+import { TouchableOpacity } from "react-native"
 import LeftArrowIcon from "../icons/LeftArrowIcon"
+import { router } from "expo-router"
 
 
 type Props = {
@@ -8,12 +9,28 @@ type Props = {
 }
 
 export default function BackButton({ onClick, className }: Props) {
+  /**
+   * Handler untuk tombol Back.
+   */
+  const handleClick = () => {
+    // Jika ada prop handler yang diberikan, jalankan.
+    if (onClick) {
+      onClick()
+    }
+
+    // Jika tidak ada, kembali ke halaman sebelumnya.
+    else {
+      router.back()
+    }
+  }
+
+
   return (
     <TouchableOpacity
       activeOpacity={0.5}
-      onPress={onClick}
-      className={`flex-row justify-center items-center py-[6px] pb-[12px] rounded-full absolute ${className}`}>
-      <LeftArrowIcon width={32} height={32} fill="white" />
+      onPress={handleClick}
+      className={`flex-row justify-center items-center py-[6px] pb-[6px] rounded-full absolute z-20 ${className}`}>
+      <LeftArrowIcon width={32} height={32} fill="white"/>
     </TouchableOpacity>
   )
 }
